Update implementation to use LazyMountableExtension
