+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 arrow = "1.1.3"
 kotlin = "1.6.21"
-kotest = "5.4.1"
+kotest = "5.5.0"
 
 [libraries]
 kotest-runner-junit5 = { module = "io.kotest:kotest-runner-junit5", version.ref = "kotest" }
