+++ kotest-assertions-arrow-fx-coroutines/src/commonMain/kotlin/io/kotest/assertions/arrow/fx/coroutines/ProjectTestResource.kt
@@ -3,6 +3,7 @@ package io.kotest.assertions.arrow.fx.coroutines
 import arrow.fx.coroutines.Resource
 import arrow.fx.coroutines.continuations.ResourceScope
 import arrow.fx.coroutines.continuations.resource
+import io.kotest.core.extensions.LazyMaterialized
 import io.kotest.core.listeners.ProjectListener
 
 /**
@@ -19,18 +20,17 @@ import io.kotest.core.listeners.ProjectListener
  *
  * class MySpec : StringSpec({
  *   "my test" {
- *     val dataSource: HikariDataSource = ProjectConfig.hikari()
- *     val dataSource2: HikariDataSource = ProjectConfig.hikari.get()
+ *     val dataSource: HikariDataSource = ProjectConfig.hikari.get()
  *   }
  * })
  * ```
  */
 public class ProjectTestResource<A> private constructor(
-  public override val resource: Resource<A>,
-  private val default: DefaultTestResource<A>
-) : ProjectListener, TestResource<A> by default {
+  public val resource: Resource<A>,
+  private val default: ResourceLazyMaterialized<A>
+) : ProjectListener, LazyMaterialized<A> by default {
 
-  public constructor(resource: Resource<A>) : this(resource, DefaultTestResource(resource))
+  public constructor(resource: Resource<A>) : this(resource, ResourceLazyMaterialized(resource))
   public constructor(block: suspend ResourceScope.() -> Resource<A>) : this(resource { block().bind() })
 
   override suspend fun beforeProject() {
