+++ kotest-assertions-arrow-fx-coroutines/src/commonTest/kotlin/io/kotest/assertions/arrow/fx/coroutines/ResourceExtensionSpecTest.kt
@@ -20,8 +20,8 @@ class ResourceExtensionSpecTest : FunSpec({
 
   afterSpec {
     sub.isReleased shouldBe true
-    shouldThrow<TestResource.AlreadyClosedException> {
-      res()
+    shouldThrow<IllegalStateException> {
+      res.get()
     }
   }
 })
